Use `cached` to cache `Middleware` requests
This enables caching support for `index` requests using `cached`.

In comparison to `memoize`, `cached` has a few advantages for our use-case:

- `memoize` doesn't support async functions. See https://github.com/dermesser/memoize/issues/27
- Return types in `memoize` need to be `Clone`, which is not the case for `Error` and it's not trivial to change that, because upstream errors like `serde_json::Error` and `std::io::Error` are not `Clone`. There are ways around it, (e.g. by using `String` as error types or implementing a `newtype`, but let's see if we can avoid that.

`cached` supports these properties out of the box.
The only caveat is, that `self` arguments are not supported, which is why I moved the `HashMap` generation
into a separate function (`index_inner`).