Add basic CI/CD pipeline
We are reaching a point where testing changes become more important.
As such, I propose to add a basic GitHub Workflow for testing the crates.

The additional changes are required to get through some of the checks.
I've excluded `terraphim_ai` for now, as this will be a separate pull request.

Also, we need to fix the benchmarks and tests.

It is okay if the build is failing; that is to be expected for the moment.
The goal is to have basic checks going forward.