Fix server start
Blocks until the server is started to fix tests.

Using `ctor` gives me a segfault on macOS.
I am not sure if it should be used in this case.
As per the [ctor README.md](https://github.com/mmastrac/rust-ctor):

> The code that runs in the ctor and dtor functions should be careful to limit
> itself to libc functions and code that does not rely on Rust's stdlib
> services.

That's why I vote for keeping it simple and manually calling the server
initialization at the beginning of each test.